--- a/gatsby-config.ts
+++ b/gatsby-config.ts
@@ -10,6 +10,13 @@ const config: GatsbyConfig = {
   plugins: [
     "gatsby-plugin-emotion",
     {
+      resolve: "gatsby-source-filesystem",
+      options: {
+        name: "assets",
+        path: `${__dirname}/src/assets`,
+      },
+    },
+    {
       resolve: "gatsby-plugin-alias-imports",
       options: {
         alias: {
@@ -19,30 +26,6 @@ const config: GatsbyConfig = {
           "@utils": path.resolve(__dirname, "src/utils"),
           "@types": path.resolve(__dirname, "src/types"),
         },
-      },
-    },
-    {
-      resolve: "gatsby-omni-font-loader",
-      options: {
-        enableListener: true,
-        preconnect: [
-          "https://fonts.gstatic.com",
-          "https://fonts.googleapis.com",
-        ],
-        web: [
-          {
-            name: "DM Serif Text",
-            file: "https://fonts.googleapis.com/css2?family=DM+Serif+Text:wght@400;500;700&display=swap",
-          },
-          {
-            name: "Overpass",
-            file: "https://fonts.googleapis.com/css2?family=Overpass:wght@400;600;800&display=swap",
-          },
-          {
-            name: "Inconsolata",
-            file: "https://fonts.googleapis.com/css2?family=Inconsolata:wght@400;700&display=swap",
-          },
-        ],
       },
     },
   ],