--- a/gatsby-ssr.tsx
+++ b/gatsby-ssr.tsx
@@ -1,3 +1,4 @@
+import "firebase/firestore"
 import type { GatsbySSR } from "gatsby"
 import React from "react"
 import { Layout } from "./src/components/Layout"