--- a/src/pages/media.tsx
+++ b/src/pages/media.tsx
@@ -1,22 +1,17 @@
-import { db } from "@app/db"
-import { LinkWithDisplay } from "@app/nimatullo-types"
+import { useDB } from "@app/hooks"
 import { Accordion } from "@components/scaffold/Accordion"
 import { ExternalAnchor } from "@components/scaffold/ExternalAnchor"
 import { ListItem } from "@components/scaffold/List"
 import { PageIntro } from "@components/scaffold/PageIntro"
-import { PageProps } from "gatsby"
 import React from "react"
 
-interface MediaPageProps extends PageProps {
-  serverData: { playlists: LinkWithDisplay[]; links: LinkWithDisplay[] }
-}
-
-const MediaPage = ({ serverData }: MediaPageProps) => {
-  const { playlists, links } = serverData
+const MediaPage = () => {
+  const { data: links, loading: linksLoading } = useDB("links")
+  const { data: playlists, loading: playlistsLoading } = useDB("playlists")
 
   return (
     <React.Fragment>
-      <PageIntro header="Links">
+      <PageIntro header="Links" loading={linksLoading}>
         <ul>
           {links.map((link) => (
             <ListItem key={link.url}>
@@ -26,7 +21,7 @@ const MediaPage = ({ serverData }: MediaPageProps) => {
         </ul>
       </PageIntro>
 
-      <PageIntro header="Playlists">
+      <PageIntro header="Playlists" loading={playlistsLoading}>
         {playlists.map((playlist) => (
           <Accordion key={playlist.url} title={playlist.title}>
             <iframe src={playlist.url} width="100%" height="500" />
@@ -35,15 +30,6 @@ const MediaPage = ({ serverData }: MediaPageProps) => {
       </PageIntro>
     </React.Fragment>
   )
-}
-
-export async function getServerData() {
-  const [playlists, links] = await Promise.all([
-    db.playlists.all(),
-    db.links.all(),
-  ])
-
-  return { props: { playlists, links } }
 }
 
 export default MediaPage