Improvements to loading performance of the site. SSR was a good experiment and I've acknowledged previously that SSG is probably the best method for a personal site like this. So I've reverted back to it and it's fixed the page navigation bug.

Also the background gifs are now webm files and are imported through graphql. These two changes took the Lighthouse performance score from 70 to 100. major bag alert ‼️

| Before | After |
|--------|--------|
| ![image](https://github.com/user-attachments/assets/fd5f2e87-5641-4239-8586-3fa518b2b400) | ![image](https://github.com/user-attachments/assets/1d11b5b3-bf56-4223-9b70-a5868db37604) | 

more specifically took the LCP from 1.3s to